pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "jiuwenclaw"
optional-dependencies = { dev = ["pyinstaller>=6.0.0"] }
version = "0.1.5"
description = "JiuwenClaw"
requires-python = ">=3.11,<3.14"
dependencies = [
    "chromadb>=1.5.0",
    "ruamel.yaml>=0.18.0",
    "lark-oapi>=1.5.3",
    "loguru>=0.7.3",
    "openjiuwen==0.1.7",
    "pip==26.0.1",
    "pgvector>=0.4.2",
    "websockets>=12.0",
    "psutil>=7.2.2",
    "aiosqlite>=0.22.1",
    "croniter>=2.0.0",
    "dingtalk-stream>=0.24.3",
    "python-socks>=2.8.1",
    "greenlet>=3.3.2",
]

[project.scripts]
jiuwenclaw-app = "jiuwenclaw.app:main"
jiuwenclaw-web = "jiuwenclaw.app_web:main"
jiuwenclaw-start = "jiuwenclaw.start_services:main"
jiuwenclaw-init = "jiuwenclaw.init_workspace:main"

[tool.setuptools]
# 入口脚本现在是 jiuwenclaw 包内的一部分，不需要 py-modules
py-modules = []

[tool.setuptools.packages.find]
where = ["."]
include = ["jiuwenclaw*"]

[tool.setuptools.package-data]
jiuwenclaw = [
    "web/dist/**/*",
    "web/dist/*",
    # config / env 模板放在 resources 目录
    "resources/**/*",
    "resources/*",
    "resources/.env.template",
    "workspace/**/*",
    "workspace/*",
]

[tool.setuptools.exclude-package-data]
jiuwenclaw = ["web/node_modules/**", "web/node_modules/*"]
